% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atkin_1.R
\name{atkin_list}
\alias{atkin_list}
\title{Crible d'atkin}
\usage{
atkin_list(n)
}
\arguments{
\item{n}{Limite supérieure pour notre crible}
}
\value{
La fonction va renvoyer la list des nombres premiers inférieurs à la limite
}
\description{
Crible d'atkin
}
\examples{
atkin_list(50) # Renvoie 2, 3, 5, 7, 11, 13, 17, 19, 23, 29, 31, 37, 41, 43, 47
}
