% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atkin_1.R
\name{atkin_prime}
\alias{atkin_prime}
\title{crible d'atkinit}
\usage{
atkin_prime(limit)
}
\arguments{
\item{n}{Limite supérieure pour notre crible}
}
\value{
La fonction va renvoyer le premier nombres premiers supérieurs à la limite supérieure
}
\description{
crible d'atkinit
}
