% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atkin_1.R
\name{create_hash}
\alias{create_hash}
\title{Créer une table de hachage optimisée avec un nombre premier}
\usage{
create_hash(limit)
}
\arguments{
\item{n}{limite supérieure pour la taille de la table, la taille finale sera le premier nombre premier > limit}
}
\value{
une liste contenant la table (liste), sa taille et la fonction de hachage
}
\description{
Créer une table de hachage optimisée avec un nombre premier
}
\examples{
my_hash <- create_hash(100)
print(my_hash$size) # Affiche le premier nombre premier > 100
}
