% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atkin_1.R
\name{hash_get}
\alias{hash_get}
\title{Récupérer un élément dans la table de hackage}
\usage{
hash_get(h, key)
}
\arguments{
\item{h}{la table de hackage créée avec create_hash()}

\item{key}{la clé numérique}
}
\value{
La valeur stockée ou Null si non trouvée
}
\description{
Récupérer un élément dans la table de hackage
}
