% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atkin_1.R
\name{hash_insert}
\alias{hash_insert}
\title{Insert an element into the Hash Table}
\usage{
hash_insert(h, key, value)
}
\arguments{
\item{h}{The hash table created with create_hash().}

\item{key}{Numeric key.}

\item{value}{The value to store.}
}
\value{
The updated hash table.
}
\description{
Insert an element into the Hash Table
}
\examples{
h <- create_hash(50)
h <- hash_insert(h, 123, "Patient Data")
}
