% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atkin_1.R
\name{is_prime_atkin}
\alias{is_prime_atkin}
\title{Check if a number is prime using Sieve of Atkin}
\usage{
is_prime_atkin(n)
}
\arguments{
\item{n}{The integer to check.}
}
\value{
Logical TRUE if prime, FALSE otherwise.
}
\description{
Check if a number is prime using Sieve of Atkin
}
\examples{
is_prime_atkin(17) # TRUE
is_prime_atkin(20) # FALSE
}
