% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atkin_1.R
\name{pam_rsa_atkin}
\alias{pam_rsa_atkin}
\title{Genere les paramètres d'une clé RSA Simulée
Simule l'étape de generation de cle RSA pour l 'anonymisation des données médicales
En utilisant le crible d'atkin pour trouver les nombres  premiers p et q}
\usage{
pam_rsa_atkin(n)
}
\arguments{
\item{n}{Taille minimale que p et q doivent atteindre}
}
\value{
une liste contenant les paramètres (p,q et n)
}
\description{
Genere les paramètres d'une clé RSA Simulée
Simule l'étape de generation de cle RSA pour l 'anonymisation des données médicales
En utilisant le crible d'atkin pour trouver les nombres  premiers p et q
}
