% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atkin_1.R
\name{prime_factors_atkin}
\alias{prime_factors_atkin}
\title{Crible d'atkin}
\usage{
prime_factors_atkin(n)
}
\arguments{
\item{n}{un entier supérieur à 1}
}
\value{
un vecteur des facteurs premiers de N
}
\description{
Crible d'atkin
}
\examples{
prime_factors_atkin(100) # Renvoie 2, 2, 5, 5
}
