% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atkin_1.R
\name{rsa_decrypt_column}
\alias{rsa_decrypt_column}
\title{Decrypt a column of IDs in a DataFrame}
\usage{
rsa_decrypt_column(df, column_name, priv_key)
}
\arguments{
\item{df}{Dataframe with anonymized data.}

\item{column_name}{Name of the column to decrypt.}

\item{priv_key}{Private key from rsa_full_process.}
}
\value{
Dataframe with the restored column.
}
\description{
Reverses the anonymization process to retrieve original IDs.
}
\examples{
# Assuming data_med is already encrypted
# data_med <- rsa_decrypt_column(data_med, "ID", keys$private_key)
}
