% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atkin_1.R
\name{rsa_decrypt_csv}
\alias{rsa_decrypt_csv}
\title{Decrypt a CSV file column using RSA}
\usage{
rsa_decrypt_csv(chemin_1, column_name, chemin_2, private_key)
}
\arguments{
\item{chemin_1}{Path to the encrypted input CSV file.}

\item{column_name}{Name of the column to decrypt.}

\item{chemin_2}{Path where the decrypted CSV will be saved.}

\item{private_key}{The RSA private key from rsa_full_process.}
}
\description{
Reads an encrypted CSV file, decrypts a specific column using an RSA private key,
and saves the restored data to a new CSV file.
}
\examples{
# keys <- rsa_full_process(1000)
# rsa_decrypt_csv("patients_anon.csv", "ID", "patients_restored.csv", keys$private_key)
}
