% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atkin_1.R
\name{rsa_encrypt_column}
\alias{rsa_encrypt_column}
\title{Anonymiser une colonne d'identifiants dans un DataFrame}
\usage{
rsa_encrypt_column(df, column_name, pub_key)
}
\arguments{
\item{df}{Le dataframe contenant les données.}

\item{column_name}{Le nom de la colonne à chiffrer (ex: "ID").}

\item{pub_key}{La clé publique issue de rsa_full_process.}
}
\value{
Le dataframe avec la colonne transformée.
}
\description{
Applique le chiffrement RSA sur chaque ligne d'une colonne spécifique.
Les résultats sont stockés en chaînes de caractères pour préserver la précision.
}
\examples{
data_med <- data.frame(ID = c(101, 102, 103), Note = c("A", "B", "A"))
keys <- rsa_full_process(1000)
data_med <- rsa_encrypt_column(data_med, "ID", keys$public_key)
}
