% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atkin_1.R
\name{rsa_full_process}
\alias{rsa_full_process}
\alias{rsa_encrypt}
\alias{rsa_decrypt}
\title{RSA Key Generation, Encryption, and Decryption}
\usage{
rsa_full_process(n_atkin)

rsa_encrypt(message_int, pub_key)

rsa_decrypt(cipher_int, priv_key)
}
\arguments{
\item{n_atkin}{Minimum size for the prime number search.}

\item{message_int}{An integer (or bigz) to encrypt.}

\item{pub_key}{Public key (list containing e and n).}

\item{cipher_int}{The encrypted integer.}

\item{priv_key}{Private key (list containing d and n).}
}
\value{
A list containing the public/private keys and parameters p, q.
}
\description{
Simulates the RSA key generation step for medical data anonymization
using the Sieve of Atkin, then provides tools for encryption.
}
\examples{
# 1. Generate system keys
my_rsa <- rsa_full_process(1000)

# 2. Encrypt a patient ID
anon_id <- rsa_encrypt(123, my_rsa$public_key)
print(anon_id)

# 3. Decrypt to find original
original_id <- rsa_decrypt(anon_id, my_rsa$private_key)
print(as.character(original_id))

}
