% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atkin_1.R
\name{atkin_list}
\alias{atkin_list}
\title{Sieve of Atkin (List generation)}
\usage{
atkin_list(n)
}
\arguments{
\item{n}{Upper limit for the sieve.}
}
\value{
A vector of prime numbers up to n.
}
\description{
Sieve of Atkin (List generation)
}
\examples{
atkin_list(50) # Returns 2, 3, 5, ..., 47
}
