% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atkin_1.R
\name{atkin_list}
\alias{atkin_list}
\title{Crible d'atkin}
\usage{
atkin_list(n)
}
\arguments{
\item{n}{Limite supérieure pour notre crible}
}
\value{
La fonction va renvoyer la list des nombres premiers inférieurs à la limite
}
\description{
Crible d'atkin
}
