% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atkin_1.R
\name{atkin_prime}
\alias{atkin_prime}
\title{Sieve of Atkin (Find next prime)}
\usage{
atkin_prime(limit)
}
\arguments{
\item{limit}{The upper bound for the search.}
}
\value{
The first prime number strictly greater than the limit.
}
\description{
Sieve of Atkin (Find next prime)
}
\examples{
atkin_prime(10) # Returns 11
atkin_prime(20) # Returns 23
}
