% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atkin_1.R
\name{hash_get}
\alias{hash_get}
\title{Retrieve an element from the Hash Table}
\usage{
hash_get(h, key)
}
\arguments{
\item{h}{The hash table.}

\item{key}{Numeric key.}
}
\value{
The stored value or NULL if not found.
}
\description{
Retrieve an element from the Hash Table
}
\examples{
h <- create_hash(50)
h <- hash_insert(h, 99, "Test Value")
hash_get(h, 99)
}
