% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atkin_1.R
\name{hash_insert}
\alias{hash_insert}
\title{Insérer un élément dans la table de hackage}
\usage{
hash_insert(h, key, value)
}
\arguments{
\item{h}{la table de hackage créée avec create_hash()}

\item{key}{la clé numérique}

\item{value}{la valeur à stocker}
}
\value{
la table mise à jour
}
\description{
Insérer un élément dans la table de hackage
}
