% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atkin_1.R
\name{is_prime_atkin}
\alias{is_prime_atkin}
\title{Crible d'atkin}
\usage{
is_prime_atkin(n)
}
\arguments{
\item{n}{Limite supérieure pour notre crible}
}
\value{
True si le nombre est premier
}
\description{
Crible d'atkin
}
\examples{
is_prime_atkin(17) # TRUE
is_prime_atkin(20) # FALSE
}
