% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atkin_1.R
\name{prime_factors_atkin}
\alias{prime_factors_atkin}
\title{Prime Factorization using Sieve of Atkin}
\usage{
prime_factors_atkin(n)
}
\arguments{
\item{n}{An integer greater than 1.}
}
\value{
A vector of prime factors.
}
\description{
Prime Factorization using Sieve of Atkin
}
\examples{
prime_factors_atkin(100) # Returns 2, 2, 5, 5
}
