% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atkin_1.R
\name{rsa_decrypt_column}
\alias{rsa_decrypt_column}
\title{Déchiffrer une colonne d'identifiants dans un DataFrame}
\usage{
rsa_decrypt_column(df, column_name, priv_key)
}
\arguments{
\item{df}{Le dataframe contenant les données anonymisées.}

\item{column_name}{Le nom de la colonne à déchiffrer.}

\item{priv_key}{La clé privée issue de rsa_full_process.}
}
\value{
Le dataframe avec la colonne restaurée.
}
\description{
Inverse le processus d'anonymisation pour retrouver les identifiants originaux.
}
