% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atkin_1.R
\name{rsa_full_process}
\alias{rsa_full_process}
\alias{rsa_encrypt}
\alias{rsa_decrypt}
\title{Génération de clés, Chiffrement et Déchiffrement RSA}
\usage{
rsa_full_process(n_atkin)

rsa_encrypt(message_int, pub_key)

rsa_decrypt(cipher_int, priv_key)
}
\arguments{
\item{n_atkin}{Taille minimale pour la recherche des nombres premiers.}

\item{message_int}{Un entier (ou bigz) à chiffrer}

\item{pub_key}{La clé publique (liste e, n)}

\item{cipher_int}{L'entier chiffré}

\item{priv_key}{La clé privée (liste d, n)}
}
\value{
Une liste contenant les clés et les paramètres p, q.
}
\description{
Simule l'étape de génération de clé RSA pour l'anonymisation des données médicales
en utilisant le crible d'Atkin, puis fournit les outils de cryptage.
}
\examples{
# 1. Générer le système
mon_rsa <- rsa_full_process(1000)

# 2. Chiffrer un identifiant patient
id_anonyme <- rsa_encrypt(123, mon_rsa$public_key)
print(id_anonyme)

# 3. Déchiffrer pour retrouver l'original
id_origine <- rsa_decrypt(id_anonyme, mon_rsa$private_key)
print(as.character(id_origine))

}
